requirements.txt
transformers>=4.36.0
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
chess>=1.9.0
python-chess>=1.9.0
requests>=2.28.0
huggingface_hub>=0.19.0
gradio>=4.0.0
streamlit
accelerate>=0.26.0
groq
python-dotenv
